+++ birdhousebuilder/recipe/nginx/__init__.py
@@ -3,17 +3,28 @@
 """Recipe nginx"""
 
 import os
+import pwd
 import stat
+from shutil import copy2
 from uuid import uuid4
 from mako.template import Template
+import logging
 
 import zc.buildout
-from birdhousebuilder.recipe import conda, supervisor
-
-import logging
-logger = logging.getLogger(__name__)
+import zc.recipe.deployment
+from zc.recipe.deployment import Configuration
+from zc.recipe.deployment import make_dir
+import birdhousebuilder.recipe.conda
+from birdhousebuilder.recipe import supervisor
 
 templ_config = Template(filename=os.path.join(os.path.dirname(__file__), "nginx.conf"))
+templ_cmd = Template(
+    '${conda_prefix}/sbin/nginx -p ${prefix} -c ${etc_prefix}/nginx/nginx.conf -g "daemon off;"')
+
+def make_dirs(name, user, mode=0o755):
+    etc_uid, etc_gid = pwd.getpwnam(user)[2:4]
+    created = []
+    make_dir(name, etc_uid, etc_gid, mode, created)
 
 def generate_cert(out, org, org_unit, hostname):
     """
@@ -43,13 +54,14 @@ def generate_cert(out, org, org_unit, hostname):
         crypto.dump_certificate(crypto.FILETYPE_PEM, cert))
         open(out, "at").write(
         crypto.dump_privatekey(crypto.FILETYPE_PEM, k))
-
-        os.chmod(out, stat.S_IRUSR|stat.S_IWUSR)
+        # TODO: fix permissions  ... nginx is run by unpriviledged user.
+        #os.chmod(out, stat.S_IRUSR|stat.S_IWUSR)
     except:
-        logger.exception("Certificate generation has failed!")
+        print("Certificate generation has failed!")
         return False
     else:
         return True
+
 
 class Recipe(object):
     """This recipe is used by zc.buildout"""
@@ -58,56 +70,83 @@ class Recipe(object):
         self.buildout, self.name, self.options = buildout, name, options
         b_options = buildout['buildout']
 
-        self.prefix = self.options.get('prefix', conda.prefix())
-        self.options['prefix'] = self.prefix
+        self.options['name'] = self.options.get('name', self.name)
+        self.name = self.options['name']
+
+        self.logger = logging.getLogger(name)
+
+        # deployment layout
+        def add_section(section_name, options):
+            if section_name in buildout._raw:
+                raise KeyError("already in buildout", section_name)
+            buildout._raw[section_name] = options
+            buildout[section_name] # cause it to be added to the working parts
+            
+        self.deployment_name = self.name + "-nginx-deployment"
+        self.deployment = zc.recipe.deployment.Install(buildout, self.deployment_name, {
+            'name': "nginx",
+            'prefix': self.options['prefix'],
+            'user': self.options['user'],
+            'etc-user': self.options['etc-user']})
+        add_section(self.deployment_name, self.deployment.options)
+        
+        self.options['etc-prefix'] = self.options['etc_prefix'] = self.deployment.options['etc-prefix']
+        self.options['var-prefix'] = self.options['var_prefix'] = self.deployment.options['var-prefix']
+        self.options['etc-directory'] = self.options['etc_directory'] = self.deployment.options['etc-directory']
+        self.options['lib-directory'] = self.options['lib_directory'] = self.deployment.options['lib-directory']
+        self.options['log-directory'] = self.options['log_directory'] = self.deployment.options['log-directory']
+        self.options['run-directory'] = self.options['run_directory'] = self.deployment.options['run-directory']
+        self.options['cache-directory'] = self.options['cache_directory'] = self.deployment.options['cache-directory']
+        self.prefix = self.options['prefix']
+
+        # conda environment path
+        self.options['env'] = self.options.get('env', '')
+        self.options['pkgs'] = self.options.get('pkgs', 'nginx openssl pyopenssl cryptography')
+        self.options['channels'] = self.options.get('channels', 'defaults birdhouse')
+        self.conda = birdhousebuilder.recipe.conda.Recipe(self.buildout, self.name, {
+            'env': self.options['env'],
+            'pkgs': self.options['pkgs'],
+            'channels': self.options['channels'] })
+        self.options['conda-prefix'] = self.options['conda_prefix'] = self.conda.options['prefix']
+
+        # config options     
         self.options['hostname'] = self.options.get('hostname', 'localhost')
-        self.options['user'] = self.options.get('user', '')
-        self.options['worker_processes'] = self.options.get('worker_processes', '1')
-        self.options['keepalive_timeout'] = self.options.get('keepalive_timeout', '5s')
+        self.options['worker-processes'] =self.options['worker_processes'] = self.options.get('worker-processes', '1')
+        self.options['keepalive-timeout'] = self.options['keepalive_timeout'] = self.options.get('keepalive-timeout', '5s')
         self.options['sendfile'] = self.options.get('sendfile', 'off')
         self.options['organization'] = self.options.get('organization', 'Birdhouse')
-        self.options['organization_unit'] = self.options.get('organization_unit', 'Demo')
+        self.options['organization-unit'] = self.options.get('organization-unit', 'Demo')
 
         self.input = options.get('input')
-        self.options['sites'] = self.options.get('sites', name)
-        self.sites = self.options['sites']
+
+        # make nginx dirs
+        for dirname in ['client', 'fastcgi', 'proxy', 'scgi', 'uwsgi']:
+            make_dirs(os.path.join(self.options['etc-directory'], dirname), self.options['etc-user'], mode=0o750)
+        
 
     def install(self, update=False):
         installed = []
-        installed += list(self.install_nginx(update))
+        if not update:
+            installed += list(self.deployment.install())
+        installed += list(self.conda.install(update))
         installed += list(self.install_cert(update))
         installed += list(self.install_config(update))
-        installed += list(self.setup_service(update))
+        installed += list(self.install_supervisor(update))
         installed += list(self.install_sites(update))
         return installed
 
-    def install_nginx(self, update):
-        script = conda.Recipe(
-            self.buildout,
-            self.name,
-            {'pkgs': 'nginx openssl pyopenssl cryptography'})
-
-        conda.makedirs( os.path.join(self.prefix, 'etc', 'nginx') )
-        conda.makedirs( os.path.join(self.prefix, 'var', 'cache', 'nginx') )
-        conda.makedirs( os.path.join(self.prefix, 'var', 'log', 'nginx') )
-
-        if update:
-            return script.update()
-        else:
-            return script.install()
-
     def install_cert(self, update):
-        certfile = os.path.join(self.prefix, 'etc', 'nginx', 'cert.pem')
+        certfile = os.path.join(self.options['etc-directory'], 'cert.pem')
         if update:
             # Skip cert generation on update mode
-	    return []
-	elif os.path.isfile(certfile):
+            return []
+        elif os.path.isfile(certfile):
             # Skip cert generation if file already exists.
             return []
         elif generate_cert(
                 out=certfile,
                 org=self.options.get('organization'),
-                org_unit=self.options.get('organization_unit'),
+                org_unit=self.options.get('organization-unit'),
                 hostname=self.options.get('hostname')):
             return []
         else:
@@ -117,64 +156,44 @@ class Recipe(object):
         """
         install nginx main config file
         """
-        result = templ_config.render(**self.options)
-
-        output = os.path.join(self.prefix, 'etc', 'nginx', 'nginx.conf')
-        conda.makedirs(os.path.dirname(output))
-        
+        text = templ_config.render(**self.options)
+        config = Configuration(self.buildout, 'nginx.conf', {
+            'deployment': self.deployment_name,
+            'text': text})
+        # copy additional files
         try:
-            os.remove(output)
-        except OSError:
+            copy2(os.path.join(os.path.dirname(__file__), "mime.types"), self.options['etc-directory'])
+        except:
             pass
+        return [config.install()]
 
-        with open(output, 'wt') as fp:
-            fp.write(result)
-        return [output]
-
-    def setup_service(self, update):
+    def install_supervisor(self, update):
         # for nginx only set chmod_user in supervisor!
         script = supervisor.Recipe(
             self.buildout,
-            self.name,
-            {'chown': self.options.get('user', ''),
+            self.name+'-nginx',
+            {'prefix': self.options['prefix'],
+             'user': self.options['user'],
+             'etc-user': self.options['etc-user'],
              'program': 'nginx',
-             'command': '%s/sbin/nginx -p %s -c %s/etc/nginx/nginx.conf -g "daemon off;"' % (self.prefix, self.prefix, self.prefix),
-             'directory': '%s/sbin' % (self.prefix),
+             'command': templ_cmd.render(**self.options),
+             'directory': '%s/sbin' % (self.options['conda-prefix']),
              })
         return script.install(update)
 
     def install_sites(self, update):
         templ_sites = Template(filename=self.input)
-        result = templ_sites.render(**self.options)
+        text = templ_sites.render(**self.options)
+        config = Configuration(self.buildout, self.name + '.conf', {
+            'deployment': self.deployment_name,
+            'directory': os.path.join(self.options['etc-directory'], 'conf.d'),
+            'text': text})
+        return [config.install()]
 
-        output = os.path.join(self.prefix, 'etc', 'nginx', 'conf.d', self.sites + '.conf')
-        conda.makedirs(os.path.dirname(output))
-        
-        try:
-            os.remove(output)
-        except OSError:
-            pass
-
-        with open(output, 'wt') as fp:
-            fp.write(result)
-        return [output]
-
-    def remove_start_stop(self):
-        output = os.path.join(self.prefix, 'etc', 'init.d', 'nginx')
-        
-        try:
-            os.remove(output)
-        except OSError:
-            pass
-        return [output]
-    
     def update(self):
         return self.install(update=True)
     
     def upgrade(self):
-        # clean up things from previous versions
-        # TODO: this is not the correct way to do it
-        #self.remove_start_stop()
         pass
 
 def uninstall(name, options):